requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
django-rest-knox==4.2.0
django-cors-headers==3.14.0
requests==2.28.2
